requirements.txt
numpy==1.26.0
opencv-python==4.8.1.78
Pillow==10.0.1
pypng==0.20220715.0
qrcode==7.4.2
typing_extensions==4.8.0
